Sync WebEnter updates as of 5/11

+++ Cloud Enter/Epi.Web.EF/EntitySurveyResponseDao.cs
@@ -2270,12 +2270,13 @@ namespace Epi.Web.EF
             using (var Context = DataObjectFactory.CreateContext())
             {
 
-                ResponseXml Response = Context.ResponseXmls.First(x => x.ResponseId == Id);
+                ResponseXml Response = Context.ResponseXmls.FirstOrDefault(x => x.ResponseId == Id);
 
-
-                Context.ResponseXmls.DeleteObject(Response);
-                Context.SaveChanges();
-
+                if (Response != null)
+                {
+	                Context.ResponseXmls.DeleteObject(Response);
+	                Context.SaveChanges();
+                }
                 //Update Status
                 //var Query = from response in Context.SurveyResponses
                 //            where response.ResponseId == Id
